requirements.txt
google-genai>=1.0.0
supabase>=2.4.0
python-dotenv>=1.0.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0